refactor(analyzers): merge vision.go into claude.go as single unified client
Consolidate the Claude analyzer into one file with one Client that
implements both BannerAnalyzer (text) and VisionAnalyzer (screenshots).

- Merge vision.go into claude.go, delete vision.go
- Extract shared API call logic into doVisionRequest/doVisionRequestRaw
- Consolidate two init() registrations into one (registers both "claude"
  and "claude-vision" with the same factory)
- Add ReadTerminalOutput to VisionAnalyzer interface
